Refactor endpoint trace graph v2 helpers
